refactor(remediate): extract the outcome vocabulary to outcome.ts (module size)
The self-guardrail blocked WP1's run.ts at the large-file threshold —
the gate dogfooding itself, same as the 4.3.6 ledger-render split. The
outcome taxonomy, envelope, result, and option shapes move to
src/remediate/outcome.ts; run.ts re-exports them, so consumers keep one
import surface.
