+++ src/remediate/run.ts
@@ -22,165 +22,31 @@
  * dxkit-authored prompts only. Landing (the standing PR) is composed by the
  * CLI/workflow layer on top of this runner's outcome.
  */
-import type { AnalysisTrustContext } from '../analysis-trust';
-import { runCorrectnessFloor, type CorrectnessFloorResult } from '../analyzers/correctness/run';
-import {
-  attributeFloorFailures,
-  type AttributedFloorFailure,
-  type FloorBaseCheck,
-} from '../analyzers/correctness/attribution';
+import { runCorrectnessFloor } from '../analyzers/correctness/run';
+import { attributeFloorFailures, type FloorBaseCheck } from '../analyzers/correctness/attribution';
 import { detectActiveLanguages } from '../languages';
 import { renderRemediateLedger } from './ledger-render';
-import { guardrailVerdictFor, toFloorBaseChecks, type GuardrailGateResult } from '../lanes/verify';
-import { resolveModelSetting, type AgentDriver, type AgentRunResult } from './driver';
+import { guardrailVerdictFor, toFloorBaseChecks } from '../lanes/verify';
+import { resolveModelSetting, type AgentRunResult } from './driver';
 import { AGENT_DRIVERS, knownDriverIds } from './registry';
 import type { RemediateTask } from './tasks';
-import { resolveDispatchedTask, type DispatchOverrides } from './dispatch';
+import { resolveDispatchedTask } from './dispatch';
 import { resumePromptNote } from './resume';
 import { realGit } from './git-ops';
-import {
-  evaluateScoreHinge,
-  healthHingeScores,
-  type HingeEvidence,
-  type HingeScores,
-} from './score-hinge';
-import type { RemediateConfig } from './config';
+import { evaluateScoreHinge, healthHingeScores } from './score-hinge';
+import type { AgentEnvelope, RemediateResult, RemediateRunOptions } from './outcome';
 
-export type RemediateOutcome =
-  | 'verified' // diff produced, floor net-new-clean, guardrail PASSED — ready to land
-  | 'no-op' // agent ran TO COMPLETION, no diff (nothing to fix)
-  | 'floor-red' // diff breaks the net-new floor — never lands
-  | 'guardrail-red' // guardrail blocked, refused, or could not run — never lands
-  | 'score-red' // the task's score hinge did not hold (goal not met) — never lands
-  | 'budget-exhausted' // a cap hit; partial diff (salvage policy decides its fate)
-  | 'agent-never-ran' // CLI/auth/env failure — infra, not a code outcome
-  | 'agent-failed' // the run errored after starting and produced no committed change
-  | 'sweep-failed' // agent committed work but leftovers could not be swept
-  | 'refused'; // trust/config refusal, disclosed
-
-export interface AgentEnvelope {
-  readonly driver: string;
-  /** Driver-native model argument + how it was chosen (ledger disclosure). */
-  readonly model: string;
-  readonly modelSource: 'auto-tier' | 'pinned-tier' | 'pinned-native';
-  readonly modelWarning?: string;
-  /** Concrete id the run reported, or absent ("not reported by driver"). */
-  readonly resolvedModelId?: string;
-  /** The agent CLI build that executed the run (run provenance), when the
-   *  driver could probe it. */
-  readonly cliVersion?: string;
-  /**
-   * Which auth path the run used: `api-key` = the runner injected a declared
-   * credential (billed API spend); `subscription` = no credential injected,
-   * the CLI's own stored login applied. Under subscription auth a reported
-   * cost is a NOTIONAL API-equivalent, not billed spend — the ledger labels
-   * it so a benchmark table never reads as a bill.
-   */
-  readonly auth: 'api-key' | 'subscription';
-  readonly turns?: number;
-  readonly costUsd?: number;
-  /** Driver-reported failure (an error after the run started) — disclosed
-   *  even when committed work verifies clean. */
-  readonly failure?: string;
-  readonly budget: {
-    readonly maxTurns: number;
-    readonly maxMinutes: number;
-    readonly maxUsd: number;
-  };
-  /** Caps the driver cannot enforce — disclosed, never silent. */
-  readonly unenforceableCaps: readonly string[];
-}
-
-export interface RemediateResult {
-  readonly outcome: RemediateOutcome;
-  readonly task?: RemediateTask['id'];
-  readonly note?: string;
-  readonly envelope?: AgentEnvelope;
-  /** True when the run was cut short (wall-clock or turn cap) — the ledger
-   *  must say the task was budget-bounded, not finished. */
-  readonly partial?: boolean;
-  readonly floor?: CorrectnessFloorResult;
-  readonly floorAttribution?: readonly AttributedFloorFailure[];
-  readonly guardrailVerdict?: string;
-  /** Score-hinge evidence (tasks that declare one): the entry vs post-agent
-   *  dimension scores the land decision was made on. Present on success AND
-   *  failure — the ledger shows the delta either way. */
-  readonly scoreHinge?: HingeEvidence;
-  /** HEAD before/after the agent — the commit range a lander pushes. */
-  readonly baseHead?: string;
-  readonly head?: string;
-  /** Dispatch-campaign disclosure (E3): who fired it, the verbatim custom
-   *  prompt (when the `custom` task ran), and any clamped overrides. In the
-   *  ledger = in the PR body, so the reviewer sees exactly what was asked. */
-  readonly dispatch?: {
-    readonly actor?: string;
-    readonly prompt?: string;
-    readonly clamped: readonly string[];
-  };
-  /** Present when this run CONTINUED a prior budget-bounded attempt
-   *  (resume-from-salvage) — disclosed in the ledger/PR body. */
-  readonly resume?: { readonly attempt: number };
-  /** The last lines of the agent's captured output — JOB-LOG evidence for a
-   *  non-clean outcome (the run page must be diagnosable without reading
-   *  source). Never rendered into the ledger / PR body. */
-  readonly transcriptTail?: string;
-  /** The verification ledger — PR body / job summary markdown. */
-  readonly ledger: string;
-}
-
-export interface RemediateGit {
-  head(): string;
-  /** Commit uncommitted leftovers (excluding dxkit runtime state); returns
-   *  an error string when the sweep commit failed. */
-  sweepLeftovers(): string | undefined;
-  /** Any commits in base..HEAD? */
-  hasDiff(baseHead: string): boolean;
-}
-
-export interface RemediateRunOptions {
-  readonly cwd: string;
-  /** REQUIRED typed trust context — the runner executes an agent CLI. */
-  readonly trust: AnalysisTrustContext;
-  readonly taskId: string;
-  readonly config: RemediateConfig;
-  /** Credentials for the driver (CI: from repo secrets). Local default is
-   *  empty — the claude-code driver then runs subscription-mode. */
-  readonly agentEnv?: Readonly<Record<string, string>>;
-  /** Injected for tests. */
-  readonly drivers?: readonly AgentDriver[];
-  readonly git?: RemediateGit;
-  readonly runFloor?: () => CorrectnessFloorResult;
-  readonly runGuardrail?: () => Promise<GuardrailGateResult>;
-  /** Injected for tests: replaces the health-score probe behind a task's
-   *  score hinge. */
-  readonly hingeScores?: (hinge: NonNullable<RemediateTask['scoreHinge']>) => Promise<HingeScores>;
-  /** Progress hook — called as each phase begins so the CLI layer can emit
-   *  log groups + heartbeats (a 35-minute silent step is indistinguishable
-   *  from a hang; "working" must be observable from the job log). */
-  readonly onPhase?: (phase: RemediatePhase) => void;
-  /** Dispatch-campaign overrides (E2/E3), read from env by the CLI layer.
-   *  Carries the custom task's prompt, the clamp disclosures, and the
-   *  dispatcher for the ledger. */
-  readonly dispatch?: DispatchOverrides;
-  /** Pre-captured entry floor (resume path): the CLI snapshots the PRISTINE
-   *  default tree BEFORE checking out a salvage branch, so attribution stays
-   *  anchored to the original base — a broken partial reads NET-NEW. */
-  readonly entryFloor?: CorrectnessFloorResult;
-  /** Present when this run continues a prior budget-bounded attempt. */
-  readonly resume?: { readonly attempt: number };
-}
-
-/** The runner's observable phases, in order. */
-export type RemediatePhase =
-  | 'entry-floor'
-  | 'agent'
-  | 'sweep'
-  | 'verify-floor'
-  | 'guardrail'
-  | 'score-hinge';
-
-// The ledger renderer lives in `./ledger-render` (module-size split);
-// re-exported so consumers keep one import surface.
+// The outcome vocabulary lives in `./outcome` and the ledger renderer in
+// `./ledger-render` (module-size splits); both are re-exported so consumers
+// keep one import surface.
+export type {
+  AgentEnvelope,
+  RemediateGit,
+  RemediateOutcome,
+  RemediatePhase,
+  RemediateResult,
+  RemediateRunOptions,
+} from './outcome';
 export { renderRemediateLedger };
 
 export async function runRemediateTask(opts: RemediateRunOptions): Promise<RemediateResult> {
